pyproject.toml
[project]
name = "memex"
version = "0.1.0"
description = "Personal knowledge base with MCP server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]",
    "httpx",
    "pyyaml",
    "python-dotenv",
    "rank-bm25",
]

[project.optional-dependencies]
semantic = ["openai"]
viewer = ["mistune"]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]

[project.scripts]
memex = "server.main:run"
